Switched from series to origin_* fields.
Series is now split into origin_title, origin_type and
origin_medium fields (eg. "Evangelion", "anime", "tv"). During the
import we approximate if something is anime or not based on
whether it has a series and if it's Japanese; otherwise we have no
idea (and leave it blank.)

Also contains additional tweaks for artist names (sometimes
inappropriately capturing "Vocal_" as part of the name).